forge fmt

+++ src/EthernautDAO/NoName/Contract.sol
@@ -2,12 +2,12 @@
 pragma solidity ^0.8.13;
 
 contract Contract {
-    uint public lastXDigits;
-    uint public mod;
+    uint256 public lastXDigits;
+    uint256 public mod;
     bool public done;
     address public winner;
 
-    constructor(uint digits, uint m) {
+    constructor(uint256 digits, uint256 m) {
         lastXDigits = digits;
         mod = m;
         done = false;
@@ -15,7 +15,7 @@ contract Contract {
 
     function cantCallMe() public {
         require(done == false, "Already done");
-        uint res = block.number % mod;
+        uint256 res = block.number % mod;
         require(res == lastXDigits, "Can't call me !");
         winner = msg.sender;
         done = true;
